deepfake-detection-and-misinterpretation-of-ai/backend/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
python-multipart==0.0.12
opencv-python==4.10.0.84
numpy==2.1.3
pillow==11.0.0
torch==2.5.1
transformers==4.46.3
safetensors==0.4.5
